+++ tools/daemon/list.do.pl
@@ -21,7 +21,6 @@
 # License, or (at your option) any later version.
 #
 # TheToolsProject is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 # General Public License for more details.
 #