+++ tools/libexec/perl/TTP/RunnerDaemon.pm
@@ -106,7 +106,7 @@ my $Const = {
 			'etc/daemons',
 			'daemons'
 		],
-		sufix => '.json'
+		suffix => '.json'
 	},
 	# how to find executable
 	exeFinder => {